🐞 [Shaders]: Allow nullptr to reset active shader
When using custom shaders there was no easy way to re-bind the default.
Setting the shader to nullptr will now cause the renderer to fallback
on to using the default provided shaders. This allows custom shaders to
be bound when rendering and then unbound when the default shader is
needed again. This fixed an issue where once a custom shader had been
set, it could not be correctly rebound or unbound.