🐞 [Sprite]: Do not use text shader for rendering
When rendering sprites, if no custom shader is attached to the sprite,
the default behaviour is to use the currently active shader. However,
the shader used for font rendering is not compatible due to its use
of the R channel for sampling. Instead, if the text shader is attached,
this change requests use of the engine's default sprite shader.