🐞 [Camera]: Fix camera zoom
When applying camera zoom, the positional offset was incorrectly being
adjusted. This would result in the view being shifted too far over.
This led to a kind-of offset zoom where the view would zoom, but be
offsetted from it's original look at position. Thanks to @olihewi for
reporting the issue.