🚧 [Renderer]: Improve resize behaviour
Improvements when attempting to scale a window to a fullscreen view.
It will now resize and reposition any viewports set when the screen
resolution does not match the requested resolution. The requested
resolution is the size set when the game launches. Currently, this does
not support window resizing.

* This feature is considered experimental and needs more testing

Note: An additional improvement might also be a callback event
to forward to the client when the window is resized or the game enters
fullscreen modes.