✨ [Texture]: Allow user to specify UV wrapping mode
Wrapping texture sample co-ordinates should be defined by the
developer as defining a default is problematic. CLAMP stops
textures from wrapping and REPEAT causes noise when filtering
is applied at the edges of textures. The default now is CLAMP,
with the option to change it on a per texture basis.