✨ [Game]: Support for fixed timesteps
This commit reintroduces the Unity style approach to updates. The
current code-base uses an XNA style system, where `update` is
always rendered with fixed deltas and `render` is variable. However,
this can be confusing, resulting in the update function becoming
the primary source of truth for the game.

The Unity approach introduces a second update step instead. To
implement this change in ASGE the existing `update` function has
been converted to a variable time-step function (indeterminate
amount of time between updates) and a new function named
`fixedUpdate` has been added, which as the name implies is a fixed
delta only function. It's currently virtual and not pure to a)
prevent breakages of existing code-bases and b) because use of it
is not essential.

The `render` function remains variable and the `update` function
is always only ever called once per drawn frame, unlike the new
`fixedUpdate` function which attempts to catch-up when it falls
behind. To prevent this from hanging the window and blocking
changes from being drawn, a default MAX_FRAMETIME value of 200ms
has been added. If the fixed updates take longer than 200ms, a
frame will always be rendered and the frame delta set to 200ms.

**It is important to note, that this may cause slow down when
performance is terrible, but should at least allow the user to
see the occasional frame and what's going on. Ideally, the fixed
updates should not take longer than the desired frametime to
help avoid these issues.**

Finally, users should use `frame_delta` for variable delta timings
and `fixed_delta` for the fixed timestep amount. At present the max
frametime can not be adjusted, but should be fairly trivial to expose
to the end-user should the need arise!

John Austin also provides a superb introduction to the Unity timestep
and why using it correctly is so important:
https://johnaustin.io/articles/2019/fix-your-unity-timestep

✨ Viva la timestep! ✨