✨ [Renderer]: Support for MSAA
Adds support for MSAA and will check the hardware to ensure it's
supported. If the sample size requested is not available the
engine will clamp the value to the maximum supported samples.

MSAA helps with edges and involves taking a number of texture
samples before combining the result to generate the final pixel
output. Think, linear filtering. The number of samples comes with a
cost to performance and memory usage. Whilst the default is an
aggressive 32 SAMPLES. Users of ASGE are free to alter it by
adjusting the MSAA setting when starting the game.

Note: the MSAA setting can not be altered once the window is
created and does not apply to nearest neighbour filtering.