💥 [Input] Fix input registrations
Note: This is an API breaking change

This commit fixes an issue where handlers could only be
successfully removed in the same order they were added. This was
due to the use of the vector's size as the handler id. If a
callback function was removed the size would change and thus all
the id's of any items that followed the removed index would be
invalid.

Instead, ASGE now makes use of the stduuid library and each event
handler added is assigned an unique id. This id, along with the
event type and the callback function are stored in an unordered_map
with the id as the key and a tuple for the type and function.

Now when registering a new event, rather then returning the UUID, it
will be converted to a string instead, which makes storage of it
simple and easy. This string can be used at a later date to
unregister the handler without effecting any of the existing
callbacks that are registered.