💫 [Sprites]: Add debug texture support
To help with debugging of sprites without texture, this change
automatically allocates a transparent debug texture to all newly
constructed sprites. This prevents the rendering from crashing
when a texture has not been attached to a sprite. It can also be
used for rendering simple blocks of colours i.e. backdrops.

The texture is transparent and pink by default and scaled with the
currently attached viewport. This is a delibrate attempt to make it
easier to see.