💫 [Sprites]: Support for texture attach flags
These new flags are designed to control how a texture is attached
to a sprite. Currently, when attaching a new texture to a sprite, it
will automatically reset the width/height and UV coordinates set on
the existing sprite. However, it may be desirable to have more granular
control over this process. These new attach flags are an attempt to
allow this to happen. They can be set to retain the sprites properties
or to allow them to be reset, depending on the developers preferences.