✨ [Renderer]: Support texture rendering
These changes allow textures to be manually rendered without having
to attach them to a sprite object first. However, in reality, it
just takes the texture and assigns it to a temporary sprite before
adding it to the render batches, ready to be drawn. This change is
mostly one of convenience and semantics as developers should be free
to store "textures" as they see fit. A good example is a tiled object
which does not store individual positional data.