✨ [Tiles]: Tile support
Add support for Tiles to the engine. Tiles are semantically
different to sprites and thus are easier to convey using their own
data structure. Unlike a sprite, they do not store any information
regarding their positioning. Instead this is left to the developer
to provide these details when rendering them. A tile map is a great
example of when this is useful and can help avoid confusion that
happens when attempting to use sprites with their own positional
data as tiles.