♻️  [Refactor]: Remove static GameSettings
This is a fairly big change to the code-base due to the way game
settings were being accessed throughout the engine. Instead, the
data will be stored within objects where they are most relevant.
This has led to most of them to be moved in to the Renderer,
however, some have been placed in to the Game itself.

GameSettings were never really intended to be used like they were.
Instead they're were designed to be a simple struct that a
launcher could generate via a UI and then pass on to the constructor
of the game. This brings it more in line with this original idea.