✨ [Resolutions]: Rewrite of viewport and resolution scaling
A particularly large re-write of the scaling and viewport code has
taken place. A new Resolution data structure has been added. Its
job is to store all the relevant resolution data such as window,
desktop and game dimensions.

As part of this refactor/re-write and to help support correct
window scaling the idea or concept of a design resolution has been
included. This design resolution known as the base resolution, is
the size of the game world and is independent from the pixel space.

Now when running the game in different resolutions the base can be
used to ensure scaling maintains the correct ratios, leaving objects
to remain in their correct proportions and to be aligned correctly
on this screen. Well, technically this depends on the new resolution
policy that's also been added. If you prefer it is possible to "fit"
the viewports to the window, but this will lead to scaling issues.

It's recommended to review the new Resolution class for more info
and guidance on the available policies.

The hope is that these changes now allow for flexible resizing of
the game window, without the developer having to invest too much
time and thought in to. Want it to scale and maintain its aspect
ratio? Just set the policy to "maintain" and be done with it 😃