✨ [Text]: SDF Rendering
Experimental support for SDF rendering based on the msdfgen library.
Currently, when fonts are loaded an atlas texture is still generated,
but instead of using the freetype library, msdfgen is used to load the
glyphs and generate MSDF's for them. At the moment the textures are
able to load and are working, but there are still a number of areas for
improvement. These include:

1. Establishing a good distance factor
2. Loading a pre-built atlas with CSV or JSON
3. Optimising the painfully slow generation
4. Support for different char sets
5. Scaling and correct field generation