✨ [Renderer]: Render States support
These new render states keep track of changes to the viewport or the
camera view. Now instead of flushing every time one of these changes
a snapshot is taken. This snapshot is then associated with each quad
that needs to be drawn via its batch number. Then when rendering the
quad batch, the correct snapshot can be used or applied as needed.

This allows for users to change these settings when rendering and
know that the state the renderer was in when the render request was
made will still be correct when the batches are rendered.

More work needs to be done to better optimise these changes,
especially in the sort method for the batch renderer, as the aim
would be to reduce the number of switches and batches required to
draw the user requested scenes.