✨ [GamePad]: Improve GamePad mappings
Updated the GamePad retrival function to use GLFW's specific
functions that allow buttons to be mapped correctly. As part of
these changes the user is now able to load updated bindings using
the input system and as a quality of life addition, a new version
of getGamePad has been added that will return the first attached
controller on the system.