🐞 [Renderer]: Track the loaded shader on GPU
A bug is present where a shader will not apply if its been set as
the active shader but never used. This is due to not knowing which
program on the GPU is currently active. This change tracks the
currently used program and will compare the requested shader's id
with the loaded one.